+++ code/src/nuvla/job/actions/component_image_state.py
@@ -2,11 +2,11 @@
 
 import logging
 
-from nuvla.connector.registry import (new_image_semantic_tag,
-                                      image_dict_to_str,
-                                      is_semantic_version)
-from nuvla.connector.utils import (unique_id,
-                                   utc_from_now_iso)
+from ...connector.registry import (new_image_semantic_tag,
+                                   image_dict_to_str,
+                                   is_semantic_version)
+from ...connector.utils import (unique_id,
+                                utc_from_now_iso)
 from nuvla.api.resources import Module, Callback, Notification
 from ..actions import action