+++ code/src/nuvla/job/actions/deployment_state.py
@@ -3,7 +3,7 @@
 import logging
 from datetime import datetime
 
-from nuvla.connector import docker_connector, docker_cli_connector, \
+from ...connector import docker_connector, docker_cli_connector, \
     docker_compose_cli_connector, kubernetes_cli_connector
 from nuvla.api.resources import Deployment, DeploymentParameter
 from .deployment_start import DeploymentBase, application_params_update, initialize_connector
@@ -18,6 +18,7 @@ def utcnow():
     return datetime.utcnow().isoformat('T', timespec='milliseconds') + 'Z'
 
 
+@action(action_name, True)
 class DeploymentStateJob(DeploymentBase):
 
     def get_component_state(self):
@@ -140,13 +141,3 @@ class DeploymentStateJob(DeploymentBase):
             raise ex
 
         return 0
-
-
-@action(action_name + '_10', True)
-class DeploymentStateJob10(DeploymentStateJob):
-    pass
-
-
-@action(action_name + '_60', True)
-class DeploymentStateJob60(DeploymentStateJob):
-    pass